+++ test/e2e/cluster.go
@@ -15,12 +15,8 @@ import (
 	"github.com/Azure/azure-sdk-for-go/services/storage/mgmt/2019-06-01/storage"
 	"github.com/Azure/go-autorest/autorest/azure"
 	"github.com/Azure/go-autorest/autorest/to"
-	projectv1 "github.com/openshift/api/project/v1"
-	projectclient "github.com/openshift/client-go/project/clientset/versioned"
 	appsv1 "k8s.io/api/apps/v1"
-	authorizationv1 "k8s.io/api/authorization/v1"
 	corev1 "k8s.io/api/core/v1"
-	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -36,78 +32,6 @@ import (
 const (
 	testPVCName = "e2e-test-claim"
 )
-
-type Project struct {
-	projectClient projectclient.Interface
-	cli           kubernetes.Interface
-	Name          string
-}
-
-func NewProject(
-	ctx context.Context, cli kubernetes.Interface, projectClient projectclient.Interface, name string,
-) Project {
-	p := Project{
-		projectClient: projectClient,
-		cli:           cli,
-		Name:          name,
-	}
-	createCall := p.projectClient.ProjectV1().Projects().Create
-	CreateK8sObjectWithRetry(ctx, createCall, &projectv1.Project{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: p.Name,
-		},
-	}, metav1.CreateOptions{})
-	return p
-}
-
-func (p Project) Delete(ctx context.Context) {
-	deleteCall := p.projectClient.ProjectV1().Projects().Delete
-	DeleteK8sObjectWithRetry(ctx, deleteCall, p.Name, metav1.DeleteOptions{})
-}
-
-// VerifyProjectIsReady verifies that the project and relevant resources have been created correctly and returns error otherwise
-func (p Project) Verify(ctx context.Context) error {
-	By("creating a SelfSubjectAccessReviews")
-	CreateK8sObjectWithRetry(ctx, p.cli.AuthorizationV1().SelfSubjectAccessReviews().Create,
-		&authorizationv1.SelfSubjectAccessReview{
-			Spec: authorizationv1.SelfSubjectAccessReviewSpec{
-				ResourceAttributes: &authorizationv1.ResourceAttributes{
-					Namespace: p.Name,
-					Verb:      "create",
-					Resource:  "pods",
-				},
-			},
-		}, metav1.CreateOptions{})
-
-	By("getting the relevant SA")
-	sa := GetK8sObjectWithRetry(
-		ctx, p.cli.CoreV1().ServiceAccounts(p.Name).Get, "default", metav1.GetOptions{},
-	)
-
-	if len(sa.Secrets) == 0 {
-		return fmt.Errorf("default ServiceAccount does not have secrets")
-	}
-
-	project := GetK8sObjectWithRetry(
-		ctx, p.projectClient.ProjectV1().Projects().Get, p.Name, metav1.GetOptions{},
-	)
-	_, found := project.Annotations["openshift.io/sa.scc.uid-range"]
-	if !found {
-		return fmt.Errorf("SCC annotation does not exist")
-	}
-	return nil
-}
-
-// VerifyProjectIsDeleted verifies that the project does not exist and returns error if a project exists
-// or if it encounters an error other than NotFound
-func (p Project) VerifyProjectIsDeleted(ctx context.Context) error {
-	_, err := p.projectClient.ProjectV1().Projects().Get(ctx, p.Name, metav1.GetOptions{})
-	if kerrors.IsNotFound(err) {
-		return nil
-	}
-
-	return fmt.Errorf("Project exists")
-}
 
 var _ = Describe("Cluster", Serial, func() {
 	var project Project
